The factory function introduces verbosity in code
The factory function requires the developer to keep track of the
observable, it's feeder and it's cancel function respectively.
So to simplify, I simple(but dangerously) extended the observable
to implement the IFn protocol from clojurescript. I also delete
the bit of code the could basically be reproduced with dev code.